many: refactor chooser to read user keypress
We plan to hide grub so we need a way to read the user input on boot
to select a different mode. In the real case the device vendor should
supply the magic trigger detector in the gadget snap, so we're running
a separate executable here to do that.

We have different strategies to detect the keypress: we can check for
an actual keypress, or see if the key is already pressed. This
implementation uses the latter for the Shift key, but the key still
needs to be pressed *after* the kernel starts to boot, and the actual
window is very small. We must find a better solution.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>